feat(main): Move main app functionality to separated page component
